requirements.txt
requests
BeautifulSoup4
python-dotenv